Fixed partially
It's gradienting, but the corners are flipped. This is because you want multiplier to be higher when the distance is less. So i have 2 options, the cheating one which is just switch the top and bottom, but it won't work for multiple inputs or (2) Figure out some way to flip the values