Fixed
Basically, you take the sum of all distances, then remove the current one, and then divide that by the sum. The lower the distance, the greater the numerator, the greater the multiplier. It looks like euclidean is more agressive and manhattan gives a smoother transition, but that might just be because manhatan logic is diff. I will now implement this logic on the manhattan