FOUND THE FORMULA
formula is rotating around average distance! you do this by taking difference of avg and distance and then * 2 and then add to original distance. Reason this works is the first time you add the difference will get you to the avg(because avg-dist+dist = avg) and the second time will get you ahead of avg if it was originally behind, or vise versa. This also maintains sum of multipliers=1 because the if you take the sum of all diferences from avg(With the sign, so + or -) it is 0. So if the total increase in distances is x, the total decrease is also x